% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{bcts}
\alias{bcts}
\title{Bayesian clinical trial simulation (BCTS)}
\usage{
bcts(
  n_per_arm_int = 20,
  n_per_arm_pln = 65,
  n_per_arm_max = 80,
  block.sizes = 1,
  mu_c = 0,
  sd_c = 1,
  mu_t = c(`Velusetrag 15mg` = 0.4, `Velusetrag 30mg` = 0.5),
  sd_t = c(`Velusetrag 15mg` = 1, `Velusetrag 30mg` = 1),
  prioritize_low_trteff = TRUE,
  gamma = 0.975,
  th.fut = 0.2,
  th.eff = 0.9,
  th.prom = 0.5,
  method = "mcmc",
  nsim = 1000,
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2,
  progress.bar = "text"
)
}
\arguments{
\item{mu_c}{Expected effect size for the control treatment}

\item{sd_c}{Expected standard deviation for the control treatment}

\item{mu_t}{Named vector with expected effect sizes for the active treatments}

\item{sd_t}{Named vector with expected standard deviations for the active treatments}

\item{prioritize_low_trteff}{If multiple treatments have a posterior predictive power > 'th.eff', the treatment with lowest effect size
will be selected at interim  (prioritize_low_trteff = TRUE).}

\item{gamma}{Level to declare success. For a fixed design, gamma is typically chosen as 0.975 for a one-sided type-I error rate of 2.5\%. However, an increase is usually needed because of dose selection.}

\item{th.fut}{Futility threshold for the predictive power calculated conditioned on the interim data}

\item{th.eff}{Efficacy threshold for the predictive power calculated conditioned on the interim data}

\item{th.prom}{Predictive power threshold that would trigger sample size increase}

\item{method}{Method to estimate the power. Choose "bayes" for a fully Bayesian approach or "mcmc" for a frequentist approximation.}

\item{nsim}{Number of trials to simulate (default: 1000)}

\item{num_chains}{Number of MCMC chains}

\item{n.iter}{Number of MCMC iterations for estimation}

\item{n.adapt}{Number of MCMC iterations for adaptation}

\item{perc_burnin}{How many n.iter should be reserved for burnin?}

\item{progress.bar}{Type of progress bar. Possible values are "text", "gui", and "none"}

\item{n_int}{Planned sample size at interim (control arm)}

\item{n_0_pln}{Planned sample size for the control group}

\item{n_0_max}{Maximum sample size for the control group after sample size increase}

\item{quiet}{Display messaging output}
}
\value{
An object of class "bcts"
}
\description{
The null hypothesis can be rejected when the posterior probability that $\mu_t$ - $\mu_c$ exceeds a high probability threshold,
i.e. $Pr(\mu_t - \mu_c > 0|D) > $\gamm$)$.
}
