% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{bcts_EVAL}
\alias{bcts_EVAL}
\title{Bayesian clinical trial simulation (BCTS)}
\usage{
bcts_EVAL(
  n_int = 60,
  n_pln = 20 + 65 * 2,
  n_max = 20 + 80 * 2,
  mu = c(Placebo = 0, `Velusetrag 15mg` = 0.4, `Velusetrag 30mg` = 0.5),
  sigma = c(Placebo = 1, `Velusetrag 15mg` = 1, `Velusetrag 30mg` = 1),
  trt_ref = "Placebo",
  trt_rank = c(`Velusetrag 15mg` = 1, `Velusetrag 30mg` = 2, Placebo = 3),
  prioritize_low_rank = TRUE,
  gamma = 0.98,
  eval.typeI = TRUE,
  method = "mcmc",
  th.fut = 0.2,
  th.eff = 0.9,
  th.prom = 0.5,
  nsim = 1000,
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2,
  summary.conf.level = 0.95,
  progress.bar = "text"
)
}
\arguments{
\item{n_int}{Interim sample size}

\item{n_pln}{Planned sample size}

\item{n_max}{Maximum sample size}

\item{mu}{Named vector with expected effect sizes}

\item{sigma}{Named vector with expected standard deviations}

\item{trt_ref}{Character denoting the control treatment}

\item{trt_rank}{Named vector with preference ranking for each treatment (e.g., lower doses are preferred)}

\item{prioritize_low_rank}{If multiple treatments have a posterior predictive power > 'th.eff', the treatment with lowest rank
will be selected at interim}

\item{gamma}{Level to declare success. For a fixed design, gamma is typically chosen as 0.975 for a one-sided type-I error rate of 2.5\%. However, an increase is usually needed because of dose selection.}

\item{eval.typeI}{Evaluate type-I error? (yes by default)}

\item{method}{Method to estimate the power. Choose "bayes" for a fully Bayesian approach or "mcmc" for a frequentist approximation.}

\item{th.fut}{Futility threshold for the predictive power calculated conditioned on the interim data}

\item{th.eff}{Efficacy threshold for the predictive power calculated conditioned on the interim data}

\item{th.prom}{Predictive power threshold that would trigger sample size increase}

\item{nsim}{Number of trials to simulate (default: 1000)}

\item{num_chains}{Number of MCMC chains}

\item{n.iter}{Number of MCMC iterations for estimation}

\item{n.adapt}{Number of MCMC iterations for adaptation}

\item{perc_burnin}{How many n.iter should be reserved for burnin?}

\item{summary.conf.level}{Confidence level for summary estimates}

\item{progress.bar}{Type of progress bar. Possible values are "text", "gui", and "none"}
}
\value{
A data frame with simulation results
}
\description{
Bayesian clinical trial simulation (BCTS)
}
