% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_one_interim.r
\name{bcts_one_interim}
\alias{bcts_one_interim}
\title{Bayesian clinical trial simulation with One Interim Analysis (BCTS)}
\usage{
bcts_one_interim(
  n_int,
  n_pln,
  n_max,
  mu,
  sigma,
  trt_ref,
  trt_rank = NULL,
  gamma = 0.975,
  th.fut = 0.2,
  th.eff = 0.9,
  th.prom = 0.5,
  method = "mcmc",
  nsim = 1000,
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2,
  progress.bar = "text"
)
}
\arguments{
\item{n_int}{Interim sample size}

\item{n_pln}{Planned sample size}

\item{n_max}{Maximum sample size}

\item{mu}{Named vector with expected effect sizes}

\item{sigma}{Named vector with expected standard deviations}

\item{trt_ref}{Character denoting the control treatment}

\item{trt_rank}{Named vector with preference ranking for each treatment (e.g., lower doses are preferred) when multiple treatments have a posterior predictive power > 'th.eff'}

\item{gamma}{Level to declare success. For a fixed design, gamma is typically chosen as 0.975 for a one-sided type-I error rate of 2.5\%. However, an increase is usually needed because of dose selection.}

\item{th.fut}{Futility threshold for the predictive power calculated conditioned on the interim data}

\item{th.eff}{Efficacy threshold for the predictive power calculated conditioned on the interim data}

\item{th.prom}{Predictive power threshold that would trigger sample size increase}

\item{method}{Method to estimate the power. Choose "bayes" for a fully Bayesian approach or "mcmc" for a frequentist approximation.}

\item{nsim}{Number of trials to simulate (default: 1000)}

\item{num_chains}{Number of MCMC chains}

\item{n.iter}{Number of MCMC iterations for estimation}

\item{n.adapt}{Number of MCMC iterations for adaptation}

\item{perc_burnin}{How many n.iter should be reserved for burnin?}

\item{progress.bar}{Type of progress bar. Possible values are "text", "gui", and "none"}
}
\value{
An object of class "bcts"
}
\description{
The null hypothesis can be rejected when the posterior probability that $\\mu_t$ - $\\mu_c$ exceeds a high probability threshold,
i.e. $Pr(\\mu_t - \\mu_c > 0|D) > $\\gamma$)$.
}
