% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_betabinom_sim.R
\name{check_alpha_bracketing}
\alias{check_alpha_bracketing}
\title{Check alpha bracketing for calibration of Type-I error}
\usage{
check_alpha_bracketing(
  calibrate_on,
  alpha,
  lower,
  upper,
  r_lo,
  r_hi,
  verbose = TRUE
)
}
\arguments{
\item{calibrate_on}{One of "point", "upper", or "lower"}

\item{alpha}{Target Type-I error}

\item{lower, upper}{Gamma bounds}

\item{r_lo}{Result of type1_at(lower)}

\item{r_hi}{Result of type1_at(upper)}

\item{verbose}{Print message}
}
\value{
Invisibly TRUE
}
\description{
Check alpha bracketing for calibration of Type-I error
}
