% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesNI-conj-calibrate.R
\name{check_B_cal}
\alias{check_B_cal}
\title{Validate B_cal parameter}
\usage{
check_B_cal(B_cal, min_reliable = 100, context = NULL)
}
\arguments{
\item{B_cal}{Number of simulations}

\item{min_reliable}{Minimum value considered reliable (default = 100)}

\item{context}{Optional name of the calling function for better messages}
}
\value{
An integer-validated B_cal value (invisibly)
}
\description{
Validate B_cal parameter
}
