% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{dose_selection}
\alias{dose_selection}
\title{Apply dose selection at interim}
\usage{
dose_selection(
  dat_int,
  n_pln,
  n_max,
  trt_ref,
  trt_active,
  trt_rank,
  prioritize_low_rank = TRUE,
  gamma = 0.975,
  th.fut = 0.2,
  th.eff = 0.9,
  th.prom = 0.5,
  method = "mcmc",
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2
)
}
\arguments{
\item{dat_int}{Observed data at the interim stage}

\item{n_pln}{Planned sample size}

\item{n_max}{Maximum sample size}

\item{trt_ref}{Character denoting the control treatment}

\item{trt_active}{Character vector denoting the active treatment names}

\item{trt_rank}{Named vector with preference ranking for each treatment (e.g., lower doses are preferred)}

\item{prioritize_low_rank}{If multiple treatments have a posterior predictive power > 'th.eff', the treatment with lowest rank
will be selected at interim}

\item{gamma}{Level to declare success. For a fixed design, gamma is typically chosen as 0.975 for a one-sided type-I error rate of 2.5\%. However, an increase is usually needed because of dose selection.}

\item{th.fut}{Futility threshold for the predictive power calculated conditioned on the interim data}

\item{th.eff}{Efficacy threshold for the predictive power calculated conditioned on the interim data}

\item{th.prom}{Predictive power threshold that would trigger sample size increase}

\item{method}{Method to estimate the power. Choose "bayes" for a fully Bayesian approach or "mcmc" for a frequentist approximation.}

\item{num_chains}{Number of MCMC chains}

\item{n.iter}{Number of MCMC iterations for estimation}

\item{n.adapt}{Number of MCMC iterations for adaptation}

\item{perc_burnin}{How many n.iter should be reserved for burnin?}
}
\value{
A list with PPoS and a dataframe containing information on the selected dose and final sample size
}
\description{
Apply dose selection at interim
}
