% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{eval_superiority}
\alias{eval_superiority}
\title{Evaluate the posterior probability of treatment superiority}
\usage{
eval_superiority(
  data,
  margin,
  gamma,
  method,
  num_chains = 4,
  n.adapt = 500,
  n.iter = 1000,
  perc_burnin = 0.2,
  trt_ref = "Placebo"
)
}
\arguments{
\item{data}{Trial data}

\item{margin}{Superiority margin (default: 0)}

\item{gamma}{Level to declare success (default: 0.975)}

\item{method}{Method to evaluate superiority (Choose 'mcmc' or 'bayes')}

\item{num_chains}{The number of parallel chains for the model}

\item{n.adapt}{The number of iterations for adaptation.}

\item{n.iter}{The number of iterations to monitor}

\item{perc_burnin}{The percentage of iterations to keep for burn-in}

\item{trt_ref}{Reference treatment name}
}
\value{
A dataframe with results from the evaluation
}
\description{
Superiority is established if Pr(mean(treat)-mean(control)>margin) > gamma
}
\references{
O’Hagan A, Stevens JW, Campbell MJ. Assurance in clinical trial design. Pharmaceut Statist. 2005 Jul;4(3):187–201.
}
