% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{eval_superiority_bayes}
\alias{eval_superiority_bayes}
\title{Bayesian method to evaluate superiority}
\usage{
eval_superiority_bayes(
  data,
  margin,
  gamma,
  trt_ref = "Placebo",
  num_chains,
  n.adapt,
  n.iter,
  perc_burnin
)
}
\arguments{
\item{data}{Trial data}

\item{margin}{Superiority margin (default: 0)}

\item{gamma}{Level to declare success (default: 0.975)}

\item{trt_ref}{Reference treatment}

\item{num_chains}{The number of parallel chains for the model}

\item{n.adapt}{The number of iterations for adaptation.}

\item{n.iter}{The number of iterations to monitor}

\item{perc_burnin}{The percentage of iterations to keep for burn-in}
}
\value{
A data frame with results from the evaluation
}
\description{
Bayesian method to evaluate superiority
}
