% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{eval_superiority_mcmc}
\alias{eval_superiority_mcmc}
\title{Frequentist method to evaluate superiority}
\usage{
eval_superiority_mcmc(data, margin = 0, gamma = 0.975, trt_ref = "Placebo")
}
\arguments{
\item{data}{Trial data}

\item{margin}{Superiority margin (default: 0)}

\item{gamma}{Level to declare success (default: 0.975)}

\item{trt_ref}{Reference treatment name}
}
\value{
A data frame with the results from the evaluation
}
\description{
Frequentist method to evaluate superiority
}
