% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_treatment_names}
\alias{extract_treatment_names}
\title{Extract Treatment Names}
\usage{
extract_treatment_names(mu, sigma)
}
\arguments{
\item{mu}{Named vector of expected effect sizes for each treatment group (including the control group).}

\item{sigma}{Named vector of expected standard deviations for each treatment group.}
}
\value{
A named list with unique treatment names.
}
\description{
Extract Treatment Names
}
