% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{filter_bcts_by_design}
\alias{filter_bcts_by_design}
\title{Filter a list of bcts objects based on design criteria}
\usage{
filter_bcts_by_design(
  bcts_list,
  no.looks = NULL,
  futility_threshold = NULL,
  gamma_values = NULL,
  ...
)
}
\arguments{
\item{bcts_list}{A list of objects of class bcts to be filtered.}

\item{no.looks}{An optional filter for the number of looks. Only bcts objects with this number of looks will be retained.}

\item{futility_threshold}{An optional filter for the futility threshold (e.g., 0.15, 0.20). Only bcts objects with this threshold will be retained.}

\item{gamma_values}{An optional filter for gamma values. Only bcts objects with the specified gamma values will be retained.}

\item{...}{Additional criteria for filtering the bcts objects.}
}
\value{
A filtered list of bcts objects that meet the design criteria.
}
\description{
This function takes a list of bcts objects and filters them based on user-specified design criteria.
The function returns a list of bcts objects that meet the given criteria.
}
