% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{generate_report}
\alias{generate_report}
\title{Generate a PDF Report for a bcts_results Object}
\usage{
generate_report(
  x,
  output_file = "sample_size_report.pdf",
  output_dir = getwd(),
  template_path = NULL
)
}
\arguments{
\item{x}{An object of class \code{bcts_results} containing the results to be included in the report.}

\item{output_file}{A character string specifying the output PDF file name (default: "sample_size_report.pdf").}

\item{output_dir}{A character string specifying the directory where the PDF file will be saved (default: current working directory).}

\item{template_path}{A character string specifying the path to a custom Quarto template file. If not provided, the default template in the package is used.}
}
\value{
The path to the generated PDF report.
}
\description{
Generate a PDF Report for a bcts_results Object
}
