% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{mc_error_proportion}
\alias{mc_error_proportion}
\title{Monte Carlo error for a proportion}
\usage{
mc_error_proportion(x, n, level)
}
\arguments{
\item{x}{Observed event count}

\item{n}{Total number of Monte Carlo Simulations}

\item{level}{Confidence level}
}
\description{
Monte Carlo error for a proportion
}
