% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.r
\name{N_two_sample_proportions}
\alias{N_two_sample_proportions}
\title{Sample size determination for Pearson Chi-square test}
\source{
Chow SC, Shao J Wang, Hansheng, Lokhnygina Y. Sample size calculations in clinical research. 2018.
}
\usage{
N_two_sample_proportions(
  p_treat = NA,
  p_control = NA,
  delta = 0,
  kappa = 1,
  alpha = 0.05,
  power = 0.8,
  test = "noninferiority/superiority",
  design = "parallel"
)
}
\arguments{
\item{p_treat}{Expected response probability in treated patients}

\item{p_control}{Expected response probability in control patients}

\item{delta}{Margin for deciding superiority (delta > 0) or non-inferiority (delta < 0)}

\item{kappa}{Treatment allocation ratio}

\item{alpha}{Desired level of significance}

\item{power}{Target power}

\item{test}{Type of test}

\item{design}{Study design}
}
\value{
A list
}
\description{
The problem of testing non-inferiority and superiority can be unified by the following hypotheses:
H0: p_treat - p_control <= delta
H1: p_treat - p_control > delta
}
\examples{
N_two_sample_proportions(p_treat = 0.8,
                         p_control = 0.8,
                         delta = -0.20,
                         kappa = 1,
                         alpha = 0.05,
                         power = 0.80)
}
