% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{plot.bcts}
\alias{plot.bcts}
\title{Plot simulation results}
\usage{
\method{plot}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Plot simulation results
}
