% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rct_betabinom_plot.R
\name{plot.bcts_calibration}
\alias{plot.bcts_calibration}
\title{Plot calibration trace for calibrated Bayesian NI threshold}
\usage{
\method{plot}{bcts_calibration}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link[=bcts_calibrate_betaBinom_conj]{bcts_calibrate_betaBinom_conj()}}}

\item{...}{Not used.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Visualizes the estimated Type-I error at different posterior thresholds
during the calibration procedure.
}
\examples{
# res <- bcts_calibrate_betaBinom_conj(...)  # Run separately
# plot(res)
}
