% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotFinalSampleSize.bcts}
\alias{plotFinalSampleSize.bcts}
\title{Plot final sample size}
\usage{
\method{plotFinalSampleSize}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Plot final sample size
}
