% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_results_plots.R
\name{plotFinalSampleSize.bcts_results}
\alias{plotFinalSampleSize.bcts_results}
\title{Plot final sample size}
\usage{
\method{plotFinalSampleSize}{bcts_results}(x, ...)
}
\arguments{
\item{x}{An object of class bcts_results}

\item{...}{Optional arguments}
}
\description{
Plot final sample size
}
