% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{plotFinalSampleSize}
\alias{plotFinalSampleSize}
\title{Plot the final sample size}
\usage{
plotFinalSampleSize(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot the final sample size
}
\details{
This is a generic function.
}
