% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimDecisions.bcts}
\alias{plotInterimDecisions.bcts}
\title{Plot interim decisions}
\usage{
\method{plotInterimDecisions}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Plot interim decisions
}
