% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimDecisions}
\alias{plotInterimDecisions}
\title{Plot Interim Decisions}
\usage{
plotInterimDecisions(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Interim Decisions
}
\details{
This is a generic function.
}
