% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimPPoS.bcts}
\alias{plotInterimPPoS.bcts}
\title{Plot of PPoS at Interim by Futility and Trial Outcome}
\usage{
\method{plotInterimPPoS}{bcts}(x, interim_index = NULL, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{interim_index}{An integer specifying which interim analysis to use for the plot. If \code{NULL} (default), the latest interim analysis is used.}

\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot of PPoS at Interim by Futility and Trial Outcome
}
\details{
This is a generic function.
}
