% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_results_plots.R
\name{plotInterimPPoS.bcts_results}
\alias{plotInterimPPoS.bcts_results}
\title{Plot of PPoS at Interim by Futility and Trial Outcome}
\usage{
\method{plotInterimPPoS}{bcts_results}(x, ...)
}
\arguments{
\item{x}{An object of class bcts_results}

\item{...}{Optional arguments}
}
\description{
Plot of PPoS at Interim by Futility and Trial Outcome
}
