% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimPPoS}
\alias{plotInterimPPoS}
\title{Plot of PPoS at Interim by Sample Size Decision and Trial Outcome}
\usage{
plotInterimPPoS(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot of PPoS at Interim by Sample Size Decision and Trial Outcome
}
\details{
This is a generic function.
}
