% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotPowerType1}
\alias{plotPowerType1}
\title{Plot the Relationship Between Type-I Error and Power Across Decision Thresholds (Gamma)}
\usage{
plotPowerType1(sim_list_power, sim_list_type1)
}
\arguments{
\item{sim_list_power}{A list of simulation results for power, where each element is an object containing power estimates.}

\item{sim_list_type1}{A list of simulation results for Type-I error, where each element is an object containing Type-I error estimates.}
}
\value{
A ggplot2 object representing the relationship between Type-I error and power across different decision thresholds (gamma).
}
\description{
This function creates a scatter plot to visualize the relationship between Type-I error and power
across different decision thresholds (gamma values). The input is two lists of simulation results:
one for power and one for Type-I error. The resulting plot shows how power and Type-I error
vary with respect to gamma, and optionally distinguishes between different design criteria.
}
\details{
The function expects the lists \code{sim_list_power} and \code{sim_list_type1} to be parallel, meaning each corresponding
element in the lists should refer to the same simulation setting (i.e., same gamma, number of looks, etc.).
It extracts the gamma values, power, Type-I error, and optionally other design-related parameters from
these lists and generates a scatter plot. Points in the plot are colored based on the gamma values and
shaped based on the prioritization of certain (e.g., lower) doses.
}
