% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotPPoS.bcts}
\alias{plotPPoS.bcts}
\title{Plot Probability of trial success}
\usage{
\method{plotPPoS}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Probability of trial success
}
\details{
This is a generic function.
}
