% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotPPoS}
\alias{plotPPoS}
\title{Plot Probability of trial success}
\usage{
plotPPoS(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot Probability of trial success
}
\details{
This is a generic function.
}
