% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotSampleSizeReEstimation.bcts}
\alias{plotSampleSizeReEstimation.bcts}
\title{Plot Sample Size Re-Estimation Results}
\usage{
\method{plotSampleSizeReEstimation}{bcts}(x, interim_index = NULL, ...)
}
\arguments{
\item{x}{A list object containing simulation results, including benefit estimates, predictive power, and sample size increases.
The list must include elements named \code{benefit}, \code{PPos}, and \code{ss_inc}, each of which should be a list containing matrices or data frames.}

\item{interim_index}{An integer specifying which interim analysis to use for the plot. If \code{NULL} (default), the latest interim analysis is used.}

\item{...}{Additional arguments passed to other methods.}
}
\value{
A ggplot2 object representing the plot of sample size re-estimation results.
}
\description{
This function plots the sample size re-estimation results from a simulation study. Each point on the plot
represents a simulation, with the x-axis showing the estimated benefit, the y-axis showing the predictive power,
and the color indicating the magnitude of the sample size increase.
}
\details{
The function extracts the estimated benefit and predictive power for the selected dose in each simulation and then
plots these against each other. The points are colored according to the magnitude of the sample size increase (\code{ss_inc}).
The color gradient legend is positioned at the top of the plot.
}
\author{
Thomas Debray
}
