% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotSampleSizeReEstimation}
\alias{plotSampleSizeReEstimation}
\title{Plot Sample Size Re-Estimation Results}
\usage{
plotSampleSizeReEstimation(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A ggplot2 object representing the plot of sample size re-estimation results.
}
\description{
This function plots the sample size re-estimation results from a simulation study. Each point on the plot
represents a simulation, with the x-axis showing the estimated benefit, the y-axis showing the predictive power,
and the color indicating the magnitude of the sample size increase.
}
\details{
The function extracts the estimated benefit and predictive power for the selected dose in each simulation and then
plots these against each other. The points are colored according to the magnitude of the sample size increase (\code{ss_inc}).
The color gradient legend is positioned at the top of the plot.
}
\author{
Thomas Debray
}
