% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{plotSelDose.bcts}
\alias{plotSelDose.bcts}
\title{Plot selected dose}
\usage{
\method{plotSelDose}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Plot selected dose
}
