% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{plotSelDose}
\alias{plotSelDose}
\title{Plot selected dose}
\usage{
plotSelDose(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Plot selected dose
}
\details{
This is a generic function.
}
