% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{posterior.bcts}
\alias{posterior.bcts}
\title{Plot posterior distribution of the mean outcome}
\usage{
\method{posterior}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Plot posterior distribution of the mean outcome
}
