% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{posterior}
\alias{posterior}
\title{Plot posterior distribution of the mean outcome}
\usage{
posterior(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Function to plot the posterior distribution of the mean outcome in the selected treatment arm
}
\details{
This is a generic function.
}
