% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{pow_diff_means}
\alias{pow_diff_means}
\title{Estimate the Power for the Difference between Two Means}
\usage{
pow_diff_means(
  n1,
  n2,
  N,
  mu1,
  mu2,
  margin = 0,
  sd1,
  sd2,
  alpha,
  alternative = "superiority"
)
}
\arguments{
\item{n1}{Number of observations in treatment group 1}

\item{n2}{Number of observations in treatment group 2}

\item{N}{Total sample size at final analysis (for calculating PPoS)}

\item{mu1}{Expected mean in treatment group 1}

\item{mu2}{Expected mean in treatment group 2}

\item{margin}{Margin for deciding superiority/non-inferiority}

\item{sd1}{Standard deviation of the outcome in treatment group 1}

\item{sd2}{Standard deviation of the outcome in treatment group 2}

\item{alpha}{One-sided alpha}

\item{alternative}{Alternative hypothesis, defaults to "Superiority"}
}
\value{
An object of class "fpower"
}
\description{
For superiority, H0 is defined as mu1 - mu2 <= margin.
}
