% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{power.bcts}
\alias{power.bcts}
\title{Extract the power of a bcts simulation}
\usage{
\method{power}{bcts}(x, ...)
}
\arguments{
\item{x}{An object of class bcts}

\item{...}{Optional arguments}
}
\description{
Extract the power of a bcts simulation
}
