% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{power}
\alias{power}
\title{Extract the power of a bcts simulation}
\usage{
power(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
Extract the power of a bcts simulation
}
\details{
This is a generic function.
}
