% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportions.r
\name{power_fisher}
\alias{power_fisher}
\title{Calculate the power for Fisher Test}
\usage{
power_fisher(
  n_control,
  p_treat = NA,
  p_control = NA,
  delta,
  tar = 1,
  alpha = 0.05,
  nsim = 10000
)
}
\arguments{
\item{n_control}{Number of patients in the control group}

\item{p_treat}{Expected proportion in the treated group}

\item{p_control}{Expected proportion in the control group}

\item{delta}{Margin}

\item{tar}{Treatment allocation ratio}

\item{alpha}{Alpha}

\item{nsim}{Number of simulations to conduct}
}
\value{
List with result
}
\description{
Calculate the power for Fisher Test
}
