% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.r
\name{print.optss}
\alias{print.optss}
\title{Print result from optimization}
\usage{
\method{print}{optss}(x, ...)
}
\arguments{
\item{x}{An object of class "optss"}

\item{...}{Optional arguments}
}
\description{
Print result from optimization
}
