% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bcts_app.R
\name{run_bcts_app}
\alias{run_bcts_app}
\title{Run the BCTS Shiny App}
\usage{
run_bcts_app(..., launch.browser = TRUE)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[shiny]{runApp}}, such as host or port.}

\item{launch.browser}{Logical. Should the app be launched in the system browser? Defaults to \code{TRUE}.}
}
\value{
None. This function is called for its side effect of launching the app.
}
\description{
Launches the interactive Shiny app included in the \code{bcts} package, which supports
Bayesian clinical trial simulations for both single-arm and randomized trials. This app provides
user-friendly controls to explore design parameters, power, and Type-I error.
}
\details{
This function locates the Shiny app bundled in the \code{inst/app} directory of the
\code{bcts} package and runs it using \code{shiny::runApp}. Ensure that the package was installed
with the app directory included.
}
\examples{
if (interactive()) {
  run_bcts_app()
}

}
