% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/sat_betabinom_power.R
\name{sat_betabinom_power}
\alias{sat_betabinom_power}
\title{Compute p-values for a t-distribution with Fixed Degrees of Freedom}
\usage{
sat_betabinom_power(
  B = 10000,
  p_t,
  n_t,
  M,
  threshold,
  prior = c("flat", "jeffreys", "beta"),
  a_base = 1,
  b_base = 1,
  show_progress = TRUE,
  method = c("exact", "simulate")
)

sat_betabinom_power(
  B = 10000,
  p_t,
  n_t,
  M,
  threshold,
  prior = c("flat", "jeffreys", "beta"),
  a_base = 1,
  b_base = 1,
  show_progress = TRUE,
  method = c("exact", "simulate")
)
}
\arguments{
\item{B}{Number of trial simulations (used only if \code{method = "simulate"}).}

\item{p_t}{True response probability.}

\item{n_t}{Sample size.}

\item{M}{Threshold for success (e.g. 0.6).}

\item{threshold}{Posterior probability threshold (e.g. 0.95).}

\item{prior}{Prior type: one of:
\itemize{
\item \code{"flat"} for a non-informative Beta(1,1) prior;
\item \code{"jeffreys"} for the Jeffreys prior Beta(0.5, 0.5);
\item \code{"beta"} for a custom Beta(\code{a_base}, \code{b_base}) prior.
}}

\item{a_base}{Alpha of Beta prior (only used if \code{prior = "beta"}).}

\item{b_base}{Beta of Beta prior (only used if \code{prior = "beta"}).}

\item{show_progress}{Show progress bar? (only relevant for simulation).}

\item{method}{Either \code{"simulate"} (default) or \code{"exact"}.}
}
\value{
A list with: estimate (power), mc_se, successes, B.

A list with \code{estimate} (power), \code{mc_se}, \code{successes}, and \code{B}.
}
\description{
Simulates a single-arm binomial trial with a conjugate Beta prior,
and computes the Bayesian power - the proportion of trials where the posterior
probability that the response rate exceeds a threshold \code{M} is greater than or equal
to a specified decision threshold.

Computes the Bayesian power for a single-arm binomial trial
using either simulation or exact analytic summation over all possible outcomes.
}
\examples{
sat_betabinom_power(
  B = 1000, p_t = 0.75, n_t = 35, M = 0.60,
  threshold = 0.95, prior = "flat"
)

sat_betabinom_power(
  B = 1000, p_t = 0.75, n_t = 35, M = 0.60,
  threshold = 0.95, prior = "flat", method = "simulate"
)

sat_betabinom_power(
  p_t = 0.75, n_t = 35, M = 0.60,
  threshold = 0.95, prior = "flat", method = "exact"
)

}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
