% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{sim_rct_normal}
\alias{sim_rct_normal}
\title{Simulate a randomized clinical trial}
\usage{
sim_rct_normal(n, mean, sd, trtnames)
}
\arguments{
\item{n}{Sample size}

\item{mean}{vector with the means of the treatment groups}

\item{sd}{vector with the standard deviations of the treatment groups}

\item{trtnames}{vector with treatment names}
}
\value{
A dataframe with the simulated trial data
}
\description{
Simulate a randomized clinical trial
}
\author{
Thomas Debray
}
