% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R,
%   R/bcts_power_sat_betaBinom_conj.R
\name{singlearm_beta_power}
\alias{singlearm_beta_power}
\title{Compute p-values for a t-distribution with Fixed Degrees of Freedom}
\usage{
singlearm_beta_power(
  B,
  p_t,
  n_t,
  M,
  threshold,
  prior = "flat",
  a_base = 1,
  b_base = 1,
  n_draws = 2000,
  show_progress = TRUE
)

singlearm_beta_power(
  B,
  p_t,
  n_t,
  M,
  threshold,
  prior = "flat",
  a_base = 1,
  b_base = 1,
  n_draws = 2000,
  show_progress = TRUE
)
}
\arguments{
\item{B}{Number of trial simulations.}

\item{p_t}{True response probability.}

\item{n_t}{Sample size.}

\item{M}{Threshold for success (e.g. 0.6).}

\item{threshold}{Posterior probability threshold (e.g. 0.95).}

\item{prior}{Prior type: "flat" or "beta".}

\item{a_base}{Alpha of Beta prior (only used if prior = "beta").}

\item{b_base}{Beta of Beta prior (only used if prior = "beta").}

\item{n_draws}{Number of posterior samples per trial.}

\item{show_progress}{Show progress bar?}
}
\value{
A list with: estimate (power), mc_se, successes, B.

A list with \code{estimate} (power), \code{mc_se}, \code{successes}, and \code{B}.
}
\description{
Simulates a single-arm binomial trial with a conjugate Beta prior,
and computes the Bayesian power - the proportion of trials where the posterior
probability that the response rate exceeds a threshold \code{M} is greater than or equal
to a specified decision threshold.
}
\examples{
singlearm_beta_power(
  B = 1000, p_t = 0.75, n_t = 35, M = 0.60,
  threshold = 0.95, prior = "flat", n_draws = 2000
)

}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
