% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.r
\name{uniroot}
\alias{uniroot}
\title{uniroot.integer.mod}
\usage{
uniroot(
  f,
  beta,
  lower = 2,
  upper = 500,
  step.power = 6,
  step.up = TRUE,
  pos.side = FALSE,
  maxiter = 20,
  ...
)
}
\arguments{
\item{f}{function for which a root is needed}

\item{lower}{minimum allowable root}

\item{upper}{maximum allowable root}

\item{step.power}{initial step size is 2^step.power}

\item{step.up}{if TRUE steps up from 'lower', if FALSE steps down from 'upper'}

\item{pos.side}{if TRUE finds integer, i, closest to the root such that f(i) >0}

\item{maxiter}{maximum number of iterations (default: 100)}

\item{...}{additional arguments to 'f'.}

\item{power}{target power value}
}
\value{
a list with the following elements:
root= the integer on the correct side of the root
f.root	= value of f at root  (output= main output with final power, output.test= output table with test at endpoint level)
iter	= number of times f was evaluated
table.iter = data.frame with the estimated N and power at each iteration
table.test = data.frame with the test at endpoint level for a given N and each simulation draw
}
\description{
Optimizer Uniroot integer modified from the ssanv function package https://github.com/cran/ssanv/blob/master/R/uniroot.integer.R
}
\keyword{internal}
