% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts.r
\name{bcts}
\alias{bcts}
\title{Bayesian Clinical Trial Simulation (BCTS)}
\usage{
bcts(
  n_dose_sel,
  n_ss_reest,
  n_pln,
  n_max,
  mu,
  sigma,
  trt_ref,
  trt_rank = NULL,
  alpha = 0.025,
  th.fut = 0.2,
  th.eff = 0.9,
  th.prom = 0.5,
  method = "mcmc",
  nsim = 1000,
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2,
  progress.bar = "text"
)
}
\arguments{
\item{n_dose_sel}{First interim sample size (for dose selection)}

\item{n_ss_reest}{Optional second interim sample size (for sample size re-estimation)}

\item{n_pln}{Planned sample size}

\item{n_max}{Maximum sample size}

\item{mu}{Named vector with expected effect sizes}

\item{sigma}{Named vector with expected standard deviations}

\item{trt_ref}{Character denoting the control treatment}

\item{trt_rank}{Named vector with preference ranking for each treatment (e.g., lower doses are preferred) when multiple treatments have a posterior predictive power > 'th.eff'}

\item{alpha}{Desired one-sided type-I error rate (default: 0.025)}

\item{th.fut}{Futility threshold for predictive power (default: 0.2)}

\item{th.eff}{Efficacy threshold for predictive power (default: 0.9)}

\item{th.prom}{Predictive power threshold to trigger sample size increase (default: 0.5)}

\item{method}{Method to estimate the power ("bayes" or "mcmc", default: "mcmc")}

\item{nsim}{Number of trials to simulate (default: 1000)}

\item{num_chains}{Number of MCMC chains (default: 4)}

\item{n.iter}{Total MCMC iterations (default: 5000)}

\item{n.adapt}{MCMC adaptation iterations (default: 500)}

\item{perc_burnin}{Proportion of iterations for burn-in (default: 0.2)}

\item{progress.bar}{Progress bar type ("text", "gui", "none")}
}
\value{
An object of class "bcts"
}
\description{
Simulates Bayesian clinical trials with up to two interim analyses for dose selection
and sample size re-estimation. The null hypothesis is rejected if the posterior probability
$Pr(\\mu_t - \\mu_c > 0 | D) > \\gamma$ exceeds a defined threshold.
}
\author{
Thomas Debray \email{tdebray@fromdatatowisdom.com}
}
