% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interim.r
\name{dose_selection}
\alias{dose_selection}
\title{Apply Dose Selection at Interim}
\usage{
dose_selection(
  dat_int,
  n_pln,
  trt_ref,
  trt_active,
  trt_rank = NULL,
  gamma = 0.975,
  th.fut = 0.2,
  th.eff = 0.9,
  method = "mcmc",
  num_chains = 4,
  n.iter = 5000,
  n.adapt = 500,
  perc_burnin = 0.2
)
}
\arguments{
\item{dat_int}{Observed data at the interim stage}

\item{n_pln}{Planned sample size}

\item{trt_ref}{Character denoting the control treatment}

\item{trt_active}{Character vector denoting the active treatment names}

\item{trt_rank}{Optional named vector with preference ranking for each treatment (e.g., lower doses are preferred) when multiple treatments have a posterior predictive power > 'th.eff'}

\item{gamma}{Level to declare success. For a fixed design, gamma is typically chosen as 0.975 for a one-sided type-I error rate of 2.5\%. However, an increase is usually needed because of dose selection.}

\item{th.fut}{Futility threshold for the predictive power calculated conditioned on the interim data}

\item{th.eff}{Efficacy threshold for the predictive power calculated conditioned on the interim data}

\item{method}{Method to estimate the power. Choose "bayes" for a fully Bayesian approach or "mcmc" for a frequentist approximation.}

\item{num_chains}{Number of MCMC chains}

\item{n.iter}{Number of MCMC iterations for estimation}

\item{n.adapt}{Number of MCMC iterations for adaptation}

\item{perc_burnin}{How many n.iter should be reserved for burnin?}
}
\value{
A list with PPoS and a dataframe containing information on the selected dose and final sample size
}
\description{
Apply Dose Selection at Interim
}
\examples{
# Generate interim data
mu <- c("Placebo" = 0, "Drug A" = 0.4, "Drug B" = 0.5)
sigma <- c("Placebo" = 1, "Drug A" = 1, "Drug B" = 1)
trt_ref <- "Placebo"
trt_active <- setdiff(names(mu), trt_ref)
dat_int <- sim_rct_normal(n = 20*3, mean = mu, sd = sigma)

# Frequentist approach for dose selection
dose_selection(dat_int = dat_int, n_pln = 150, trt_ref = trt_ref,
               trt_active = trt_active, gamma = 0.98)

# Bayesian approach for dose selection
\dontrun{
dose_selection(dat_int = dat_int, n_pln = 150, trt_ref = trt_ref,
               trt_active = trt_active, method = "bayes")
}
}
