% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{mc_error_proportion}
\alias{mc_error_proportion}
\title{Monte Carlo Error for Proportion Estimate}
\usage{
mc_error_proportion(x, n, level = 0.95)
}
\arguments{
\item{x}{Numeric. The observed event count.}

\item{n}{Numeric. The total number of Monte Carlo simulations.}

\item{level}{Numeric. The confidence level for the error bounds (default is 0.95).}
}
\value{
A data frame containing the observed event count, total simulations, proportion estimate, standard error, and confidence interval.
}
\description{
This function calculates the Monte Carlo error for a proportion estimate based on observed event counts.
}
\author{
Thomas Debray
}
