% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimDecisions.bcts}
\alias{plotInterimDecisions.bcts}
\title{Interim Decision Outcomes: Comparative Effect Sizes by Dose}
\usage{
\method{plotInterimDecisions}{bcts}(
  x,
  color_scheme = "dose_selection",
  interim_index = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class \code{bcts}. This object should contain the simulation results and interim decisions related to dose selection and futility.}

\item{color_scheme}{A character string specifying the coloring scheme to be used. Options are "dose_selection" and "trial_decision".
The "dose_selection" scheme colors points based on dose selection decisions ("Select dose 1", "Select dose 2", "Futility").
The "trial_decision" scheme colors points based on trial decisions ("Continue", "Expand", "Futility").}

\item{interim_index}{An integer specifying which interim analysis to plot.
If \code{NULL} (default), the first interim analysis is used.}

\item{...}{Optional arguments.}
}
\value{
A \code{ggplot2} object showing the scatter plot of estimated effect sizes for the two doses, colored according to the specified decision scheme.
}
\description{
Plot Interim Decision Outcomes for bcts Objects
}
\details{
This method visualizes the interim decision outcomes from adaptive trial simulations for objects of class \code{bcts}.
It plots the estimated effect sizes of two different doses, with points colored according to the specified decision scheme.
}
\seealso{
\code{\link{plotInterimDecisions}} for the generic function.
}
\author{
Thomas Debray
}
