% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcts_plots.r
\name{plotInterimDecisions}
\alias{plotInterimDecisions}
\title{Plot Interim Decision Outcomes}
\usage{
plotInterimDecisions(...)
}
\arguments{
\item{\dots}{Additional arguments, which are currently ignored.}
}
\value{
A \code{ggplot2} object.
}
\description{
This generic function creates a plot to visualize the interim decision outcomes from adaptive trial simulations.
The specific method used depends on the class of the input object.
}
\details{
This is a generic function.
}
